move Throwables to impl package
In some cases it would be useful in other spectator sub-projects
to keep the behavior consistent (for example #269). Also adds
comments to make sure it is clear these should not be used
outside of spectator itself.
