+++ spectator-api/src/test/java/com/netflix/spectator/impl/ThrowablesTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.spectator.api;
+package com.netflix.spectator.impl;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
